+++ clang/unittests/Tooling/RecursiveASTVisitorTests/Callbacks.cpp
@@ -403,16 +403,14 @@ void test() {
 }
 )cpp";
 
-  // TraverseUnaryOperator is not called because RecursiveASTVisitor treats
-  // individual operators as subclasses, for which it calls their Traverse
-  // methods.
   EXPECT_TRUE(visitorCallbackLogEqual(
       RecordingVisitor(ShouldTraversePostOrder::No), Code,
       R"txt(
 WalkUpFromStmt CompoundStmt
 WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromStmt UnaryOperator(-)
-WalkUpFromStmt IntegerLiteral(2)
+TraverseUnaryOperator UnaryOperator(-)
+  WalkUpFromStmt UnaryOperator(-)
+  WalkUpFromStmt IntegerLiteral(2)
 WalkUpFromStmt IntegerLiteral(3)
 )txt"));
 
@@ -420,8 +418,9 @@ WalkUpFromStmt IntegerLiteral(3)
       RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
       R"txt(
 WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromStmt UnaryOperator(-)
+TraverseUnaryOperator UnaryOperator(-)
+  WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromStmt UnaryOperator(-)
 WalkUpFromStmt IntegerLiteral(3)
 WalkUpFromStmt CompoundStmt
 )txt"));
@@ -469,20 +468,18 @@ void test() {
 }
 )cpp";
 
-  // TraverseUnaryOperator is not called because RecursiveASTVisitor treats
-  // individual operators as subclasses, for which it calls their Traverse
-  // methods.
   EXPECT_TRUE(visitorCallbackLogEqual(
       RecordingVisitor(ShouldTraversePostOrder::No), Code,
       R"txt(
 WalkUpFromStmt CompoundStmt
 WalkUpFromExpr IntegerLiteral(1)
   WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromUnaryOperator UnaryOperator(-)
-  WalkUpFromExpr UnaryOperator(-)
-    WalkUpFromStmt UnaryOperator(-)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
+TraverseUnaryOperator UnaryOperator(-)
+  WalkUpFromUnaryOperator UnaryOperator(-)
+    WalkUpFromExpr UnaryOperator(-)
+      WalkUpFromStmt UnaryOperator(-)
+  WalkUpFromExpr IntegerLiteral(2)
+    WalkUpFromStmt IntegerLiteral(2)
 WalkUpFromExpr IntegerLiteral(3)
   WalkUpFromStmt IntegerLiteral(3)
 )txt"));
@@ -492,11 +489,12 @@ WalkUpFromExpr IntegerLiteral(3)
       R"txt(
 WalkUpFromExpr IntegerLiteral(1)
   WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromUnaryOperator UnaryOperator(-)
-  WalkUpFromExpr UnaryOperator(-)
-    WalkUpFromStmt UnaryOperator(-)
+TraverseUnaryOperator UnaryOperator(-)
+  WalkUpFromExpr IntegerLiteral(2)
+    WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromUnaryOperator UnaryOperator(-)
+    WalkUpFromExpr UnaryOperator(-)
+      WalkUpFromStmt UnaryOperator(-)
 WalkUpFromExpr IntegerLiteral(3)
   WalkUpFromStmt IntegerLiteral(3)
 WalkUpFromStmt CompoundStmt
@@ -568,193 +566,6 @@ WalkUpFromStmt CompoundStmt
 )txt"));
 }
 
-TEST(RecursiveASTVisitor, StmtCallbacks_TraverseUnaryMinus) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool TraverseUnaryMinus(UnaryOperator *UO) {
-      recordCallback(__func__, UO,
-                     [&]() { RecordingVisitorBase::TraverseUnaryMinus(UO); });
-      return true;
-    }
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test() {
-  1;
-  -2;
-  3;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromStmt IntegerLiteral(1)
-TraverseUnaryMinus UnaryOperator(-)
-  WalkUpFromStmt UnaryOperator(-)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromStmt IntegerLiteral(3)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromStmt IntegerLiteral(1)
-TraverseUnaryMinus UnaryOperator(-)
-  WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromStmt UnaryOperator(-)
-WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor,
-     StmtCallbacks_TraverseUnaryMinus_WalkUpFromUnaryMinus) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool TraverseUnaryMinus(UnaryOperator *UO) {
-      recordCallback(__func__, UO,
-                     [&]() { RecordingVisitorBase::TraverseUnaryMinus(UO); });
-      return true;
-    }
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-
-    bool WalkUpFromExpr(Expr *E) {
-      recordCallback(__func__, E,
-                     [&]() { RecordingVisitorBase::WalkUpFromExpr(E); });
-      return true;
-    }
-
-    bool WalkUpFromUnaryMinus(UnaryOperator *UO) {
-      recordCallback(__func__, UO,
-                     [&]() { RecordingVisitorBase::WalkUpFromUnaryMinus(UO); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test() {
-  1;
-  -2;
-  3;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-TraverseUnaryMinus UnaryOperator(-)
-  WalkUpFromUnaryMinus UnaryOperator(-)
-    WalkUpFromExpr UnaryOperator(-)
-      WalkUpFromStmt UnaryOperator(-)
-  WalkUpFromExpr IntegerLiteral(2)
-    WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-TraverseUnaryMinus UnaryOperator(-)
-  WalkUpFromExpr IntegerLiteral(2)
-    WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromUnaryMinus UnaryOperator(-)
-    WalkUpFromExpr UnaryOperator(-)
-      WalkUpFromStmt UnaryOperator(-)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor, StmtCallbacks_WalkUpFromUnaryMinus) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-
-    bool WalkUpFromExpr(Expr *E) {
-      recordCallback(__func__, E,
-                     [&]() { RecordingVisitorBase::WalkUpFromExpr(E); });
-      return true;
-    }
-
-    bool WalkUpFromUnaryMinus(UnaryOperator *UO) {
-      recordCallback(__func__, UO,
-                     [&]() { RecordingVisitorBase::WalkUpFromUnaryMinus(UO); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test() {
-  1;
-  -2;
-  3;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromUnaryMinus UnaryOperator(-)
-  WalkUpFromExpr UnaryOperator(-)
-    WalkUpFromStmt UnaryOperator(-)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromUnaryMinus UnaryOperator(-)
-  WalkUpFromExpr UnaryOperator(-)
-    WalkUpFromStmt UnaryOperator(-)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
 TEST(RecursiveASTVisitor, StmtCallbacks_TraverseBinaryOperator) {
   class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
   public:
@@ -783,17 +594,15 @@ void test() {
 }
 )cpp";
 
-  // TraverseBinaryOperator is not called because RecursiveASTVisitor treats
-  // individual operators as subclasses, for which it calls their Traverse
-  // methods.
   EXPECT_TRUE(visitorCallbackLogEqual(
       RecordingVisitor(ShouldTraversePostOrder::No), Code,
       R"txt(
 WalkUpFromStmt CompoundStmt
 WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromStmt BinaryOperator(+)
-WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromStmt IntegerLiteral(3)
+TraverseBinaryOperator BinaryOperator(+)
+  WalkUpFromStmt BinaryOperator(+)
+  WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromStmt IntegerLiteral(3)
 WalkUpFromStmt IntegerLiteral(4)
 )txt"));
 
@@ -801,9 +610,10 @@ WalkUpFromStmt IntegerLiteral(4)
       RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
       R"txt(
 WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt BinaryOperator(+)
+TraverseBinaryOperator BinaryOperator(+)
+  WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromStmt IntegerLiteral(3)
+  WalkUpFromStmt BinaryOperator(+)
 WalkUpFromStmt IntegerLiteral(4)
 WalkUpFromStmt CompoundStmt
 )txt"));
@@ -851,22 +661,20 @@ void test() {
 }
 )cpp";
 
-  // TraverseBinaryOperator is not called because RecursiveASTVisitor treats
-  // individual operators as subclasses, for which it calls their Traverse
-  // methods.
   EXPECT_TRUE(visitorCallbackLogEqual(
       RecordingVisitor(ShouldTraversePostOrder::No), Code,
       R"txt(
 WalkUpFromStmt CompoundStmt
 WalkUpFromExpr IntegerLiteral(1)
   WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromBinaryOperator BinaryOperator(+)
-  WalkUpFromExpr BinaryOperator(+)
-    WalkUpFromStmt BinaryOperator(+)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
+TraverseBinaryOperator BinaryOperator(+)
+  WalkUpFromBinaryOperator BinaryOperator(+)
+    WalkUpFromExpr BinaryOperator(+)
+      WalkUpFromStmt BinaryOperator(+)
+  WalkUpFromExpr IntegerLiteral(2)
+    WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromExpr IntegerLiteral(3)
+    WalkUpFromStmt IntegerLiteral(3)
 WalkUpFromExpr IntegerLiteral(4)
   WalkUpFromStmt IntegerLiteral(4)
 )txt"));
@@ -876,13 +684,14 @@ WalkUpFromExpr IntegerLiteral(4)
       R"txt(
 WalkUpFromExpr IntegerLiteral(1)
   WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromBinaryOperator BinaryOperator(+)
-  WalkUpFromExpr BinaryOperator(+)
-    WalkUpFromStmt BinaryOperator(+)
+TraverseBinaryOperator BinaryOperator(+)
+  WalkUpFromExpr IntegerLiteral(2)
+    WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromExpr IntegerLiteral(3)
+    WalkUpFromStmt IntegerLiteral(3)
+  WalkUpFromBinaryOperator BinaryOperator(+)
+    WalkUpFromExpr BinaryOperator(+)
+      WalkUpFromStmt BinaryOperator(+)
 WalkUpFromExpr IntegerLiteral(4)
   WalkUpFromStmt IntegerLiteral(4)
 WalkUpFromStmt CompoundStmt
@@ -958,202 +767,6 @@ WalkUpFromStmt CompoundStmt
 )txt"));
 }
 
-TEST(RecursiveASTVisitor, StmtCallbacks_TraverseBinAdd) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool TraverseBinAdd(BinaryOperator *BO) {
-      recordCallback(__func__, BO,
-                     [&]() { RecordingVisitorBase::TraverseBinAdd(BO); });
-      return true;
-    }
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test() {
-  1;
-  2 + 3;
-  4;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAdd BinaryOperator(+)
-  WalkUpFromStmt BinaryOperator(+)
-  WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt IntegerLiteral(4)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAdd BinaryOperator(+)
-  WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(3)
-  WalkUpFromStmt BinaryOperator(+)
-WalkUpFromStmt IntegerLiteral(4)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor, StmtCallbacks_TraverseBinAdd_WalkUpFromBinAdd) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool TraverseBinAdd(BinaryOperator *BO) {
-      recordCallback(__func__, BO,
-                     [&]() { RecordingVisitorBase::TraverseBinAdd(BO); });
-      return true;
-    }
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-
-    bool WalkUpFromExpr(Expr *E) {
-      recordCallback(__func__, E,
-                     [&]() { RecordingVisitorBase::WalkUpFromExpr(E); });
-      return true;
-    }
-
-    bool WalkUpFromBinAdd(BinaryOperator *BO) {
-      recordCallback(__func__, BO,
-                     [&]() { RecordingVisitorBase::WalkUpFromBinAdd(BO); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test() {
-  1;
-  2 + 3;
-  4;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAdd BinaryOperator(+)
-  WalkUpFromBinAdd BinaryOperator(+)
-    WalkUpFromExpr BinaryOperator(+)
-      WalkUpFromStmt BinaryOperator(+)
-  WalkUpFromExpr IntegerLiteral(2)
-    WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromExpr IntegerLiteral(3)
-    WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromExpr IntegerLiteral(4)
-  WalkUpFromStmt IntegerLiteral(4)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAdd BinaryOperator(+)
-  WalkUpFromExpr IntegerLiteral(2)
-    WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromExpr IntegerLiteral(3)
-    WalkUpFromStmt IntegerLiteral(3)
-  WalkUpFromBinAdd BinaryOperator(+)
-    WalkUpFromExpr BinaryOperator(+)
-      WalkUpFromStmt BinaryOperator(+)
-WalkUpFromExpr IntegerLiteral(4)
-  WalkUpFromStmt IntegerLiteral(4)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor, StmtCallbacks_WalkUpFromBinAdd) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-
-    bool WalkUpFromExpr(Expr *E) {
-      recordCallback(__func__, E,
-                     [&]() { RecordingVisitorBase::WalkUpFromExpr(E); });
-      return true;
-    }
-
-    bool WalkUpFromBinAdd(BinaryOperator *BO) {
-      recordCallback(__func__, BO,
-                     [&]() { RecordingVisitorBase::WalkUpFromBinAdd(BO); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test() {
-  1;
-  2 + 3;
-  4;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromBinAdd BinaryOperator(+)
-  WalkUpFromExpr BinaryOperator(+)
-    WalkUpFromStmt BinaryOperator(+)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromExpr IntegerLiteral(4)
-  WalkUpFromStmt IntegerLiteral(4)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromBinAdd BinaryOperator(+)
-  WalkUpFromExpr BinaryOperator(+)
-    WalkUpFromStmt BinaryOperator(+)
-WalkUpFromExpr IntegerLiteral(4)
-  WalkUpFromStmt IntegerLiteral(4)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
 TEST(RecursiveASTVisitor, StmtCallbacks_TraverseCompoundAssignOperator) {
   class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
   public:
@@ -1182,17 +795,15 @@ void test(int a) {
 }
 )cpp";
 
-  // TraverseCompoundAssignOperator is not called because RecursiveASTVisitor
-  // treats individual operators as subclasses, for which it calls their
-  // Traverse methods.
   EXPECT_TRUE(visitorCallbackLogEqual(
       RecordingVisitor(ShouldTraversePostOrder::No), Code,
       R"txt(
 WalkUpFromStmt CompoundStmt
 WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromStmt CompoundAssignOperator(+=)
-WalkUpFromStmt DeclRefExpr(a)
-WalkUpFromStmt IntegerLiteral(2)
+TraverseCompoundAssignOperator CompoundAssignOperator(+=)
+  WalkUpFromStmt CompoundAssignOperator(+=)
+  WalkUpFromStmt DeclRefExpr(a)
+  WalkUpFromStmt IntegerLiteral(2)
 WalkUpFromStmt IntegerLiteral(3)
 )txt"));
 
@@ -1200,9 +811,10 @@ WalkUpFromStmt IntegerLiteral(3)
       RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
       R"txt(
 WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromStmt DeclRefExpr(a)
-WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromStmt CompoundAssignOperator(+=)
+TraverseCompoundAssignOperator CompoundAssignOperator(+=)
+  WalkUpFromStmt DeclRefExpr(a)
+  WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromStmt CompoundAssignOperator(+=)
 WalkUpFromStmt IntegerLiteral(3)
 WalkUpFromStmt CompoundStmt
 )txt"));
@@ -1251,22 +863,20 @@ void test(int a) {
 }
 )cpp";
 
-  // TraverseCompoundAssignOperator is not called because RecursiveASTVisitor
-  // treats individual operators as subclasses, for which it calls their
-  // Traverse methods.
   EXPECT_TRUE(visitorCallbackLogEqual(
       RecordingVisitor(ShouldTraversePostOrder::No), Code,
       R"txt(
 WalkUpFromStmt CompoundStmt
 WalkUpFromExpr IntegerLiteral(1)
   WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromCompoundAssignOperator CompoundAssignOperator(+=)
-  WalkUpFromExpr CompoundAssignOperator(+=)
-    WalkUpFromStmt CompoundAssignOperator(+=)
-WalkUpFromExpr DeclRefExpr(a)
-  WalkUpFromStmt DeclRefExpr(a)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
+TraverseCompoundAssignOperator CompoundAssignOperator(+=)
+  WalkUpFromCompoundAssignOperator CompoundAssignOperator(+=)
+    WalkUpFromExpr CompoundAssignOperator(+=)
+      WalkUpFromStmt CompoundAssignOperator(+=)
+  WalkUpFromExpr DeclRefExpr(a)
+    WalkUpFromStmt DeclRefExpr(a)
+  WalkUpFromExpr IntegerLiteral(2)
+    WalkUpFromStmt IntegerLiteral(2)
 WalkUpFromExpr IntegerLiteral(3)
   WalkUpFromStmt IntegerLiteral(3)
 )txt"));
@@ -1276,13 +886,14 @@ WalkUpFromExpr IntegerLiteral(3)
       R"txt(
 WalkUpFromExpr IntegerLiteral(1)
   WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromExpr DeclRefExpr(a)
-  WalkUpFromStmt DeclRefExpr(a)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromCompoundAssignOperator CompoundAssignOperator(+=)
-  WalkUpFromExpr CompoundAssignOperator(+=)
-    WalkUpFromStmt CompoundAssignOperator(+=)
+TraverseCompoundAssignOperator CompoundAssignOperator(+=)
+  WalkUpFromExpr DeclRefExpr(a)
+    WalkUpFromStmt DeclRefExpr(a)
+  WalkUpFromExpr IntegerLiteral(2)
+    WalkUpFromStmt IntegerLiteral(2)
+  WalkUpFromCompoundAssignOperator CompoundAssignOperator(+=)
+    WalkUpFromExpr CompoundAssignOperator(+=)
+      WalkUpFromStmt CompoundAssignOperator(+=)
 WalkUpFromExpr IntegerLiteral(3)
   WalkUpFromStmt IntegerLiteral(3)
 WalkUpFromStmt CompoundStmt
@@ -1350,207 +961,6 @@ WalkUpFromExpr DeclRefExpr(a)
 WalkUpFromExpr IntegerLiteral(2)
   WalkUpFromStmt IntegerLiteral(2)
 WalkUpFromCompoundAssignOperator CompoundAssignOperator(+=)
-  WalkUpFromExpr CompoundAssignOperator(+=)
-    WalkUpFromStmt CompoundAssignOperator(+=)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor, StmtCallbacks_TraverseBinAddAssign) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool TraverseBinAddAssign(CompoundAssignOperator *CAO) {
-      recordCallback(__func__, CAO, [&]() {
-        RecordingVisitorBase::TraverseBinAddAssign(CAO);
-      });
-      return true;
-    }
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test(int a) {
-  1;
-  a += 2;
-  3;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAddAssign CompoundAssignOperator(+=)
-  WalkUpFromStmt CompoundAssignOperator(+=)
-  WalkUpFromStmt DeclRefExpr(a)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromStmt IntegerLiteral(3)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAddAssign CompoundAssignOperator(+=)
-  WalkUpFromStmt DeclRefExpr(a)
-  WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromStmt CompoundAssignOperator(+=)
-WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor,
-     StmtCallbacks_TraverseBinAddAssign_WalkUpFromBinAddAssign) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool TraverseBinAddAssign(CompoundAssignOperator *CAO) {
-      recordCallback(__func__, CAO, [&]() {
-        RecordingVisitorBase::TraverseBinAddAssign(CAO);
-      });
-      return true;
-    }
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-
-    bool WalkUpFromExpr(Expr *E) {
-      recordCallback(__func__, E,
-                     [&]() { RecordingVisitorBase::WalkUpFromExpr(E); });
-      return true;
-    }
-
-    bool WalkUpFromBinAddAssign(CompoundAssignOperator *CAO) {
-      recordCallback(__func__, CAO, [&]() {
-        RecordingVisitorBase::WalkUpFromBinAddAssign(CAO);
-      });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test(int a) {
-  1;
-  a += 2;
-  3;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAddAssign CompoundAssignOperator(+=)
-  WalkUpFromBinAddAssign CompoundAssignOperator(+=)
-    WalkUpFromExpr CompoundAssignOperator(+=)
-      WalkUpFromStmt CompoundAssignOperator(+=)
-  WalkUpFromExpr DeclRefExpr(a)
-    WalkUpFromStmt DeclRefExpr(a)
-  WalkUpFromExpr IntegerLiteral(2)
-    WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-TraverseBinAddAssign CompoundAssignOperator(+=)
-  WalkUpFromExpr DeclRefExpr(a)
-    WalkUpFromStmt DeclRefExpr(a)
-  WalkUpFromExpr IntegerLiteral(2)
-    WalkUpFromStmt IntegerLiteral(2)
-  WalkUpFromBinAddAssign CompoundAssignOperator(+=)
-    WalkUpFromExpr CompoundAssignOperator(+=)
-      WalkUpFromStmt CompoundAssignOperator(+=)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-WalkUpFromStmt CompoundStmt
-)txt"));
-}
-
-TEST(RecursiveASTVisitor, StmtCallbacks_WalkUpFromBinAddAssign) {
-  class RecordingVisitor : public RecordingVisitorBase<RecordingVisitor> {
-  public:
-    RecordingVisitor(ShouldTraversePostOrder ShouldTraversePostOrderValue)
-        : RecordingVisitorBase(ShouldTraversePostOrderValue) {}
-
-    bool WalkUpFromStmt(Stmt *S) {
-      recordCallback(__func__, S,
-                     [&]() { RecordingVisitorBase::WalkUpFromStmt(S); });
-      return true;
-    }
-
-    bool WalkUpFromExpr(Expr *E) {
-      recordCallback(__func__, E,
-                     [&]() { RecordingVisitorBase::WalkUpFromExpr(E); });
-      return true;
-    }
-
-    bool WalkUpFromBinAddAssign(CompoundAssignOperator *CAO) {
-      recordCallback(__func__, CAO, [&]() {
-        RecordingVisitorBase::WalkUpFromBinAddAssign(CAO);
-      });
-      return true;
-    }
-  };
-
-  StringRef Code = R"cpp(
-void test(int a) {
-  1;
-  a += 2;
-  3;
-}
-)cpp";
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::No), Code,
-      R"txt(
-WalkUpFromStmt CompoundStmt
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromBinAddAssign CompoundAssignOperator(+=)
-  WalkUpFromExpr CompoundAssignOperator(+=)
-    WalkUpFromStmt CompoundAssignOperator(+=)
-WalkUpFromExpr DeclRefExpr(a)
-  WalkUpFromStmt DeclRefExpr(a)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromExpr IntegerLiteral(3)
-  WalkUpFromStmt IntegerLiteral(3)
-)txt"));
-
-  EXPECT_TRUE(visitorCallbackLogEqual(
-      RecordingVisitor(ShouldTraversePostOrder::Yes), Code,
-      R"txt(
-WalkUpFromExpr IntegerLiteral(1)
-  WalkUpFromStmt IntegerLiteral(1)
-WalkUpFromExpr DeclRefExpr(a)
-  WalkUpFromStmt DeclRefExpr(a)
-WalkUpFromExpr IntegerLiteral(2)
-  WalkUpFromStmt IntegerLiteral(2)
-WalkUpFromBinAddAssign CompoundAssignOperator(+=)
   WalkUpFromExpr CompoundAssignOperator(+=)
     WalkUpFromStmt CompoundAssignOperator(+=)
 WalkUpFromExpr IntegerLiteral(3)